Check that an SNS message is surrounded by double quotes before stripping them
Fix for [JENKINS-19337]

Stripping the double quotes is now captured in an if/else clause.

Note that I was unable to duplicate creating a double-quote-surrounded message in SNS.
Also I tried to write a test case that passed in a message surrounded by double quotes
(as the plug-in was originally coded) but this causes an earlier line of code in the
extractJsonFromPayload function to error.
So I'm not really sure how it ever worked.